Separate unit and integration testing
